refactor(broken): user login feature flag and events moved to single enum file
