+++ src/modules/domain/enums/events.enum.ts
@@ -1,5 +1,27 @@
 
-export enum EventsEnum {
+export enum QueueEventsEnum {
 	INVALID = 'INVALID',
-	NEW_CONNECTION = 'NEW_CONNECTION',
+}
+
+export enum EmitterEventsEnum {
+	DISABLE_LOGIN = 'disable_login',
+}
+
+export enum WebSocketEventsEnum {
+	// * Socket.io Events
+	CONNECT = 'connection',
+	DISCONNECT = 'disconnect',
+
+	// ? External Events
+	RECONNECT = 'reconnect',
+	EMIT = 'emit',
+	ERROR = 'error',
+
+	// ? Internal Events
+	EMIT_PRIVATE = 'emit_private',
+	BROADCAST = 'broadcast',
+	NEW_CONNECTION = 'new_connection',
+}
+export enum WebSocketRoomsEnum {
+	NEW_CONNECTIONS = 'new_connections',
 }
