+++ src/modules/events/queue/handlers/schemas/event.schema.ts
@@ -1,5 +1,5 @@
 import Joi from 'joi';
-import { EventsEnum } from '@domain/enums/events.enum';
+import { QueueEventsEnum } from '@domain/enums/events.enum';
 
 
 export interface EventSchemaInterface {
@@ -7,7 +7,7 @@ export interface EventSchemaInterface {
 	schema?: string,
 	schemaVersion?: number,
 	payload: {
-		event: EventsEnum,
+		event: QueueEventsEnum,
 		[key: string]: unknown,
 	},
 	source?: string,
@@ -18,7 +18,7 @@ const eventSchema: Joi.Schema<EventSchemaInterface> = Joi.object({
 	id: Joi.alternatives().try(Joi.number(), Joi.string()),
 	timestamp: Joi.alternatives().try(Joi.date(), Joi.string()),
 	payload: Joi.object({
-		event: Joi.string().valid(...Object.values(EventsEnum)).required(),
+		event: Joi.string().valid(...Object.values(QueueEventsEnum)).required(),
 	}).unknown(true).required(),
 }).unknown(true).required();
